+++ demos/supabase-simple-chat/lib/app_config_template.dart
@@ -1,6 +1,7 @@
+// Copy this template: `cp lib/app_config_template.dart lib/app_config.dart`
+// Edit lib/app_config.dart and enter your Supabase and PowerSync project details.
 class AppConfig {
-  static const String supabaseUrl = 'https://aaaaaaaaaaaa.supabase.co';
-  static const String supabaseAnonKey = 'eyeyeyeyeyeye';
-  static const String powersyncUrl =
-      'https://65130c9db6679a3682ba380a.powersync.journeyapps.com';
+  static const String supabaseUrl = 'https://foo.supabase.co';
+  static const String supabaseAnonKey = 'foo';
+  static const String powersyncUrl = 'https://foo.powersync.journeyapps.com';
 }