+++ demos/supabase-todolist/lib/powersync.dart
@@ -151,7 +151,8 @@ Future<String> getDatabasePath() async {
 
 Future<void> openDatabase() async {
   // Open the local database
-  db = PowerSyncDatabase(schema: schema, path: await getDatabasePath());
+  db = PowerSyncDatabase(
+      schema: schema, path: await getDatabasePath(), logger: attachedLogger);
   await db.initialize();
 
   await loadSupabase();