+++ packages/powersync_attachments_helper/example/getting_started.dart
@@ -18,13 +18,14 @@ class PhotoAttachmentQueue extends AbstractAttachmentQueue {
       : super(db: db, remoteStorage: remoteStorage);
 
   @override
-  Future<Attachment> savePhoto(String photoId, int size) async {
-    String filename = '$photoId.jpg';
+  Future<Attachment> saveFile(String fileId, int size,
+      {mediaType = 'image/jpeg'}) async {
+    String filename = '$fileId.jpg';
     Attachment photoAttachment = Attachment(
-      id: photoId,
+      id: fileId,
       filename: filename,
       state: AttachmentState.queuedUpload.index,
-      mediaType: 'image/jpeg',
+      mediaType: mediaType,
       localUri: getLocalFilePathSuffix(filename),
       size: size,
     );
@@ -33,10 +34,10 @@ class PhotoAttachmentQueue extends AbstractAttachmentQueue {
   }
 
   @override
-  Future<Attachment> deletePhoto(String photoId) async {
-    String filename = '$photoId.jpg';
+  Future<Attachment> deleteFile(String fileId) async {
+    String filename = '$fileId.jpg';
     Attachment photoAttachment = Attachment(
-        id: photoId,
+        id: fileId,
         filename: filename,
         state: AttachmentState.queuedDelete.index);
 
@@ -44,7 +45,7 @@ class PhotoAttachmentQueue extends AbstractAttachmentQueue {
   }
 
   @override
-  StreamSubscription<void> watchIds() {
+  StreamSubscription<void> watchIds({String fileExtension = 'jpg'}) {
     return db.watch('''
       SELECT photo_id FROM users
       WHERE photo_id IS NOT NULL
@@ -52,9 +53,9 @@ class PhotoAttachmentQueue extends AbstractAttachmentQueue {
       return results.map((row) => row['photo_id'] as String).toList();
     }).listen((ids) async {
       List<String> idsInQueue = await attachmentsService.getAttachmentIds();
-      for (String id in ids) {
-        await syncingService.reconcileId(id, idsInQueue);
-      }
+      List<String> relevantIds =
+          ids.where((element) => !idsInQueue.contains(element)).toList();
+      syncingService.processIds(relevantIds, fileExtension);
     });
   }
 }